✨ mbk-44: simplify kv helper functions and implement Store methods
